chore: 상태 관리 로직 리팩토링
